+++ CasualConversation/Targets/Domain/Sources/Interface/Repository/ConversationRepositoryProtocol.swift
@@ -5,8 +5,6 @@
 //  Created by Yongwoo Marco on 2022/06/28.
 //
 
-import Foundation
-
 public protocol ConversationRepositoryProtocol {
 	
 	// TODO: Conversation 추가
