+++ CasualConversation/Targets/Domain/Sources/Service/AudioService/AudioStatus.swift
@@ -6,7 +6,6 @@
 //  Copyright © 2022 pseapplications. All rights reserved.
 //
 
-import Foundation
 
 public enum AudioStatus: Int, CustomStringConvertible {
