--- CasualConversation/Targets/Domain/Tests/DomainTests.swift
@@ -1,10 +1,0 @@
-import Foundation
-import XCTest
-
-final class DomainTests: XCTestCase {
-	
-    func test_example() {
-        XCTAssertEqual("Domain", "Domain")
-    }
-	
-}
